Conflicts:
	source/test/ir/index_manager/t_bitvector.cpp

+++ source/test/ir/index_manager/t_FSIndexInput.cpp
@@ -43,24 +43,24 @@ BOOST_AUTO_TEST_CASE(read)
     // call IndexInput::readIntBySmallEndian()
     FSIndexInput fsIndexInput(fileName);
     BOOST_CHECK_EQUAL(fsIndexInput.length(), bufferSize);
-    unsigned int readIntCount = size / 3;
-    for(unsigned int i=0; i<readIntCount; ++i)
+    int readIntCount = size / 3;
+    for(int i=0; i<readIntCount; ++i)
         BOOST_CHECK_EQUAL(fsIndexInput.readIntBySmallEndian(), i);
-    BOOST_CHECK_EQUAL(fsIndexInput.getFilePointer(), readIntCount * sizeof(int));
+    BOOST_CHECK_EQUAL(fsIndexInput.getFilePointer(), static_cast<int64_t>(readIntCount * sizeof(int)));
 
     // call IndexInput::read()
-    const unsigned int newBufferSize = readIntCount * sizeof(int);
+    const int newBufferSize = readIntCount * sizeof(int);
     int* newBuffer = new int[newBufferSize];
     fsIndexInput.read(reinterpret_cast<char*>(newBuffer), newBufferSize);
-    for(unsigned int i=0; i<readIntCount; ++i)
+    for(int i=0; i<readIntCount; ++i)
         BOOST_CHECK_EQUAL(newBuffer[i], readIntCount + i);
     delete[] newBuffer;
-    BOOST_CHECK_EQUAL(fsIndexInput.getFilePointer(), 2 * readIntCount * sizeof(int));
+    BOOST_CHECK_EQUAL(fsIndexInput.getFilePointer(), static_cast<int64_t>(2 * readIntCount * sizeof(int)));
 
     // call IndexInput::readInt()
-    for(unsigned int i=0; i<(size-2*readIntCount); ++i)
+    for(int i=0; i<static_cast<int>(size-2*readIntCount); ++i)
         BOOST_CHECK_EQUAL(fsIndexInput.readInt(), toHighEnd(2*readIntCount + i));
-    BOOST_CHECK_EQUAL(fsIndexInput.getFilePointer(), bufferSize);
+    BOOST_CHECK_EQUAL(fsIndexInput.getFilePointer(), static_cast<int64_t>(bufferSize));
 }
 
 BOOST_AUTO_TEST_SUITE_END()
